+++ app/lib/pages/tabs/settings_tab.dart
@@ -494,7 +494,7 @@ class SettingsTab extends StatelessWidget {
                     padding: const EdgeInsets.only(bottom: 0),
                     children: [
                       _ButtonEntry(
-                        label: 'PRP - Peer Relay',
+                        label: t.hotspotRelayPage.title,
                         buttonLabel: t.general.open,
                         onTap: () async {
                           await context.push(() => const PrpPage());